added maven support and features
